Adicionando rota para obter post unico, rota para deletar posts e configuracao de upload para amazon s3 ou local

+++ src/controllers/PostController.js
@@ -2,6 +2,7 @@ const Post = require('../models/Post');
 const sharp = require('sharp');
 const path = require('path');
 const fs = require('fs');
+require('dotenv').config()
 
 module.exports = {
     async index(req, res) {
@@ -9,23 +10,32 @@ module.exports = {
         return res.json(posts);
     },
 
+    async indexObj(req, res) {
+        const post = await Post.findById(req.params.id);
+        return res.json(post);
+    },
+
     async store(req, res) {
-
-        const { author, place, description, hashtags } = req.body;
-        const { filename: image, key, location: url = "" } = req.file;  
         
-        let fileName = url;
+        const { author, place, description, hashtags } = req.body;
+        let fileName = "";
 
-        if (process.env.STORAGE_TYPE === 'local'){
+        if (req.file){
+            const { filename: image, key: key, location: url = "" } = req.file;  
+            fileName = url;
+        }
+
+        if ( (process.env.STORAGE_TYPE === 'local') && fileName){
 
             const [name] = image.split('.');
-            fileName = `${name}.jpg`;
+
+            fileName = `http://${process.env.APP_HOST}/files/${name}.jpg`;
     
             await sharp(req.file.path)
             .resize(500)
             .jpeg({ quality: 70 })
             .toFile(
-                path.resolve(req.file.destination, 'resized', fileName) 
+                path.resolve(req.file.destination, 'resized', name + '.jpg') 
             );
     
             fs.unlinkSync(req.file.path);
@@ -43,5 +53,15 @@ module.exports = {
         req.io.emit('post', post);
 
         return res.json(post);
-    }
+    },
+
+    async delete(req, res) {
+
+        const post = await Post.findById(req.params.id);
+        await post.remove();
+        return res.json({
+            mensagem: "Postagem deletada com sucesso",
+        });
+    },
+
 };
